+++ database/unit.h
@@ -15,7 +15,7 @@ public:
 
   struct Unit : public AbstractDatabase::AbstractEntry
   {
-    u64 army_id_ = 0;
+    i32 army_id_ = 0;
   };
 
   struct UnitReadable : public AbstractDatabase::AbstractEntryReadable
@@ -23,17 +23,17 @@ public:
     string army_txt_ = { "" };
   };
 
-  vector<u64> army_id_;
+  vector<i32> army_id_;
 
-  Units(const vector<string>& field_names, u32 id_field_position)
+  Units(const vector<string>& field_names, i32 id_field_position)
     : AbstractDatabase::AbstractDatabase(field_names, id_field_position){};
-  u64 add(const string& name);
-  u64 add(const string& name, const u64 army_id);
-  opt<UnitReadable> get_readable(const u64 id);
-  bool del(const u64 id, Units& trashbin);
-  std::optional<Unit> get(const u64 id);
-  vector<u64> get_ids_by_army(const u64 army_id);
-  u64 append(const Unit& unit);
+  i32 add(const string& name);
+  i32 add(const string& name, const i32 army_id);
+  opt<UnitReadable> get_readable(const i32 id);
+  bool del(const i32 id, Units& trashbin);
+  std::optional<Unit> get(const i32 id);
+  vector<i32> get_ids_by_army(const i32 army_id);
+  i32 append(const Unit& unit);
 
   CEREAL_LD_SV(army_id_, id_, txt_id_, curr_id_, frag_);
 };
